+++ packages/expo-manifests/ios/EXManifests/EXManifestsBaseManifest.h
@@ -26,7 +26,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (nullable NSString *)orientation;
 - (nullable NSDictionary *)experiments;
 - (nullable NSDictionary *)developer;
-- (nullable NSString *)logUrl;
 - (nullable NSString *)facebookAppId;
 - (nullable NSString *)facebookApplicationName;
 - (BOOL)facebookAutoInitEnabled;