+++ packages/expo-manifests/ios/EXManifests/EXManifestsBaseManifest.m
@@ -119,14 +119,6 @@
   return [expoGoConfig nullableDictionaryForKey:@"developer"];
 }
 
-- (nullable NSString *)logUrl {
-  NSDictionary *expoGoConfig = self.expoGoConfigRootObject;
-  if (!expoGoConfig) {
-    return nil;
-  }
-  return [expoGoConfig nullableStringForKey:@"logUrl"];
-}
-
 - (nullable NSString *)facebookAppId {
   NSDictionary *expoClientConfig = self.expoClientConfigRootObject;
   if (!expoClientConfig) {