+++ packages/expo-manifests/ios/EXManifests/EXManifestsManifest.h
@@ -56,7 +56,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (nullable NSString *)orientation;
 - (nullable NSDictionary *)experiments;
 - (nullable NSDictionary *)developer;
-- (nullable NSString *)logUrl;
 - (nullable NSString *)facebookAppId;
 - (nullable NSString *)facebookApplicationName;
 - (BOOL)facebookAutoInitEnabled;