+++ packages/expo/src/Expo.ts
@@ -1,7 +1,4 @@
 import './Expo.fx';
 
-import * as Logs from './logs/Logs';
-
-export { Logs };
 export { disableErrorHandling } from './errors/ExpoErrorManager';
 export { default as registerRootComponent } from './launch/registerRootComponent';